M3 Finished
I linked the button in welcome screen to login screen. I also deleted a
lot of files that were not need atm. Made a branch of before this so
that if we need them in the future than we can find them

P.S. I used generic login screen from Eclipse so there is A LOT of stuff
in there that I didn't code and i DON'T understand.