Implemented the Login Screen
I set up the app to follow Model-View-Presenter. I also implemented the
login screen. I removed an identifier for the src files, so when
creating a new activity in the future you might need to drag it to the
correct activities section.

There are a lot of text (.log) files that I don't know where they came
from

Fun Fact: You have to use .equals() when comparing strings or else you
might spend over an hour looking for your mistake in the only 3 lines of
code that could possible by causing your app not to work. :D